requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
ecdsa==0.19.1
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
